New observations now expire.

+++ src/main/java/edu/whimc/observationdisplayer/Observation.java
@@ -1,10 +1,13 @@
 package edu.whimc.observationdisplayer;
 
 import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.stream.Collectors;
 
+import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -14,8 +17,6 @@ import org.bukkit.inventory.ItemStack;
 import com.gmail.filoghost.holographicdisplays.api.Hologram;
 import com.gmail.filoghost.holographicdisplays.api.HologramsAPI;
 import com.gmail.filoghost.holographicdisplays.api.handler.TouchHandler;
-import com.gmail.filoghost.holographicdisplays.api.line.ItemLine;
-import com.gmail.filoghost.holographicdisplays.api.line.TextLine;
 
 import edu.whimc.observationdisplayer.utils.Utils;
 
@@ -31,51 +32,84 @@ public class Observation {
     private Location viewLoc;
     private String observation;
     private Hologram hologram;
+    private Timestamp expiration;
 
     private Observation() {}
 
-    public static void createObservation(ObservationDisplayer plugin, Player player, Location viewLoc, String observation) {
+    public static void createObservation(ObservationDisplayer plugin, Player player, Location viewLoc,
+            String observation, Timestamp expiration) {
         Timestamp timestamp = new Timestamp(System.currentTimeMillis());
-        Observation obs = new Observation(plugin, -1, timestamp, player.getName(), viewLoc, observation, true);
+        Observation obs = new Observation(plugin, -1, timestamp, player.getName(), viewLoc, observation, expiration, true);
         observations.add(obs);
     }
 
-    public static void loadObservation(ObservationDisplayer plugin, int id, Timestamp timestamp, String playerName, Location viewLoc, String observation) {
-        Observation obs = new Observation(plugin, id, timestamp, playerName, viewLoc, observation, false);
+    public static void loadObservation(ObservationDisplayer plugin, int id, Timestamp timestamp,
+            String playerName, Location viewLoc, String observation, Timestamp expiration) {
+        Observation obs = new Observation(plugin, id, timestamp, playerName, viewLoc, observation, expiration, false);
         observations.add(obs);
     }
 
-    private Observation(ObservationDisplayer plugin, int id, Timestamp timestamp, String playerName, Location viewLoc, String observation, boolean isNew) {
+    public static void scanForExpiredObservations(ObservationDisplayer plugin) {
+        Bukkit.getScheduler().scheduleSyncRepeatingTask(plugin, () -> {
+            long count = observations.stream()
+                    .filter(v -> Instant.now().isAfter(v.getExpiration().toInstant()))
+                    .collect(Collectors.toList())
+                    .stream()
+                    .peek(Observation::deleteObservation)
+                    .count();
+            if (count > 0) {
+                plugin.getQueryer().makeExpiredObservationsInactive(dbCount -> {
+                    Utils.debug("Removed " + count + " expired observation(s). (" + dbCount + ") from database");
+                });
+            }
+
+        }, 20 * 60, 20 * 60);
+    }
+
+    private Observation(ObservationDisplayer plugin, int id, Timestamp timestamp, String playerName,
+            Location viewLoc, String observation, Timestamp expiration, boolean isNew) {
         this.plugin = plugin;
         this.timestamp = timestamp;
         this.playerName = playerName;
         this.holoLoc = viewLoc.clone().add(0, 3, 0).add(viewLoc.getDirection().multiply(2));
         this.viewLoc = viewLoc;
         this.observation = observation;
+        this.expiration = expiration;
 
-        createHologram(plugin);
-
-        if (isNew) {
-            this.id = plugin.getQueryer().storeObservation(timestamp, playerName, viewLoc, observation);
-        } else {
+        if (!isNew) {
             this.id = id;
+            createHologram();
+            return;
         }
+
+        plugin.getQueryer().storeNewObservation(this, newId -> {
+            this.id = newId;
+            createHologram();
+        });
     }
 
-    private void createHologram(ObservationDisplayer plugin) {
+    private void createHologram() {
         Hologram holo = HologramsAPI.createHologram(plugin, holoLoc);
-        ItemLine signLine = holo.appendItemLine(new ItemStack(Material.OAK_SIGN));
-        TextLine observationLine = holo.appendTextLine(ChatColor.translateAlternateColorCodes('&', observation));
-        TextLine infoLine = holo.appendTextLine(ChatColor.GRAY + playerName + " - " + Utils.getDate(timestamp));
-
         ObservationClick clickListener = new ObservationClick(viewLoc);
 
-        signLine.setTouchHandler(clickListener);
-        observationLine.setTouchHandler(clickListener);
-        infoLine.setTouchHandler(clickListener);
+        holo.appendItemLine(new ItemStack(Material.OAK_SIGN))
+                .setTouchHandler(clickListener);
+        holo.appendTextLine(ChatColor.translateAlternateColorCodes('&', observation))
+                .setTouchHandler(clickListener);
+        holo.appendTextLine(ChatColor.GRAY + playerName + " - " + Utils.getDate(timestamp))
+                .setTouchHandler(clickListener);
+
+        if (expiration != null) {
+            holo.appendTextLine(ChatColor.GRAY + "Expires " + Utils.getDate(expiration))
+                    .setTouchHandler(clickListener);
+        }
 
         this.hologram = holo;
+    }
 
+    public void reRender() {
+        deleteHologramOnly();
+        createHologram();
     }
 
     private class ObservationClick implements TouchHandler {
@@ -109,36 +143,48 @@ public class Observation {
     }
 
     public Hologram getHologram() {
-        return hologram;
+        return this.hologram;
     }
 
     public String getPlayer() {
-        return playerName;
+        return this.playerName;
     }
 
     public Location getHoloLocation() {
-        return holoLoc;
+        return this.holoLoc;
     }
 
     public Location getViewLocation() {
-        return viewLoc;
+        return this.viewLoc;
     }
 
     public String getObservation() {
-        return observation;
+        return this.observation;
     }
 
     public int getId() {
-        return id;
+        return this.id;
     }
 
     public Timestamp getTimestamp() {
-        return timestamp;
+        return this.timestamp;
+    }
+
+    public Timestamp getExpiration() {
+        return this.expiration;
+    }
+
+    public boolean hasExpired() {
+        return this.timestamp != null && this.expiration.toInstant().isAfter(Instant.now());
+    }
+
+    public void setExpiration(Timestamp timestamp) {
+        this.expiration = timestamp;
     }
 
     @Override
     public String toString() {
-        String text = "&f&l" + this.observation;
+        String text = Utils.color("&f&l" + this.observation);
         if (ChatColor.stripColor(text).length() > 20) {
             text = Utils.coloredSubstring(text, 20) + "&7 . . .";
         }
@@ -148,14 +194,25 @@ public class Observation {
         this.holoLoc.getBlockY() + ", "+ this.holoLoc.getBlockZ() + "&7)";
     }
 
-    public void remove() {
-        manualRemove();
+    public void deleteAndRemoveFromDatabase() {
+        deleteAndRemoveFromDatabase(() -> {});
+    }
+
+    public void deleteAndRemoveFromDatabase(Runnable callback) {
+        plugin.getQueryer().makeSingleObservationInactive(this.id, callback);
+        deleteObservation();
+    }
+
+    public void deleteObservation() {
+        deleteHologramOnly();
         observations.remove(this);
     }
 
-    public void manualRemove() {
-        plugin.getQueryer().makeObservationInactive(this.id);
-        this.hologram.delete();
+    public void deleteHologramOnly() {
+        if (this.hologram != null) {
+            this.hologram.delete();
+            this.hologram = null;
+        }
     }
 
 }
